Implemented custom TokenObtainPairSerializer
This commit introduces a custom TokenObtainPairSerializer, MyTokenObtainPairSerializer, which extends the functionality of the TokenObtainPairSerializer from rest_framework_simplejwt.

The custom serializer adds custom claims to the JWT token and includes additional fields in the response data during token validation. These custom claims and fields provide information such as the username, email, and user role associated with the authenticated user.